api/requirements/base.txt
# Core
pytz==2021.3
Django==4.0
django-configurations==2.3.1
gunicorn==20.1.0
newrelic==7.2.4.171

# For the persistence stores
psycopg2-binary==2.9.2
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.2.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.13.0
Markdown==3.3.6
django-filter==21.1